ncaa-2026/app/pyproject.toml
[project]
name = "ncaa-bracket-2026"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "databricks-sdk>=0.30.0",
    "databricks-sql-connector>=3.0.0",
    "pydantic>=2.0.0",
    "pandas>=2.0.0",
]
